load_dtsessions_from_gcs_to_bigquery/requirements.txt
google-cloud-bigquery==2.10.0
google-cloud-storage==1.36.1